Major refactor of the network
Since blocks may have more than one previous block (even with different
output types) the previous network topology was not sufficient.

Now each block that wants to interact with the network has to ask to the
topology its senders/receivers and those will be allocated on-the-fly
lazily (before they where allocated by the scheduler, but since the
input types are not stored inside the block this is no longer possible).

Each block may have zero, one or more preceding blocks, this means that
each block does not have a single receiver. To identify the receivers
Coord is therefore insufficient. ReceiverEndpoint has been introduced,
the "address" of a receiver is the Coord of the replica + the id of the
block that is sending.

This commit also implements the multiplexer/demultiplexer of the network
channels. At the moment each replica of the same block inside each host
share the same socket (i.e. one port for each block of the job graph).
Similarly all the replicas of a block that connect to a remote replica
share the same socket (i.e. one connection for each block of the job
graph, for each remote host).