Cleanup multiplexer, add test and fix bug
Before this the topology created the demultiplexers wrongly, causing
deadlocks when a block had multiple inputs.

The demultiplexer is now unique for each block in a host, regardless on
how many ReceiverEndpoints are registered, and regardless of their input
types. This means that DemultiplexingReceiver is no longer generic over
a type parameter, but instead will spawn a
for each registered type of message.

The total number of threads for the network is therefore:
- for each block of the job graph:
  - one for the network listener (that listen for incoming connections)
  - one for each incoming connection (blocked on socket.recv())
  - one for the demultiplexer (that forwards the messages to the
    TypedDemultiplexingReceiver)
  - one for each type of incoming message (that deserialize and forward)
  - for each next block of the job graph
    - for each host that has at least one of such blocks
      - one for the multiplexer (serialize and send via the socket)