New refactor of the network topology
Each block of the job graph, inside each host, now have one
demultiplexer for each previous block (of the job graph), for each host.

This means that a block no longer has a single demux (and a single TCP
socket), but may have many of them. This was done to avoid possible
deadlocks in rare, but not impossible scenarios of heavily unbalanced
loads incoming into a block (consider a zip where 1M elements of A
arrive before the first of B, with a single channel B has to wait for
the A's to be processed, but they can't be without a B).

W.r.t the previous version, this reduces the number of threads for the
blocks with a single input (there is no need for a separate deserializer
thread), but increase the number of threads and open ports for the
blocks with multiple inputs.

Assigning the ports is not as easy as before, the number of incoming
blocks influences the number of multiplexers of a block. To solve this
problem the network topology deterministically assigns each port just
after connecting all the blocks.