Basic profiler that collects data per-thread
This profiler instruments the blocks, the network and the iterations:
by calling `get_profiler()` you can track metrics that will be stored
inside a thread-local variable that does not require locks or atomics.

The metrics are stored bucketed per time with a default resolution of
20ms (ie 50 aggregated data-points are available per second).

All those metrics are transformed in no-ops if the `profiler` feature is
not enabled (it's disabled by default). If this feature is enabled the
results are printed in json in the log.