Fix a rare race condition in StartBlock with the cache
Assume that left side has the cache. If the right side ends (with
Terminate) before the left side, the Terminate is let through but the
cache is never filled up and therefore its Terminate is never consumed
(and waited for).

At the end, it is possible that the sender of the left side fails
because it cannot send more items since the receiver exited too early.